+++ src/curve/stable/StablePair.sol
@@ -146,7 +146,7 @@ contract StablePair is ReservoirPair {
         override
         returns (uint256 amountOut)
     {
-        (uint104 lReserve0, uint104 lReserve1, uint32 lBlockTimestampLast,) = _lockAndLoad();
+        (uint256 lReserve0, uint256 lReserve1, uint32 lBlockTimestampLast,) = _lockAndLoad();
         require(amount != 0, "SP: AMOUNT_ZERO");
         uint256 amountIn;
         ERC20 tokenOut;
@@ -201,7 +201,7 @@ contract StablePair is ReservoirPair {
         uint256 lReceived = tokenOut == token0 ? lBalance1 - lReserve1 : lBalance0 - lReserve0;
         require(lReceived >= amountIn, "SP: INSUFFICIENT_AMOUNT_IN");
 
-        _updateAndUnlock(lBalance0, lBalance1, lReserve0, lReserve1, lBlockTimestampLast);
+        _updateAndUnlock(lBalance0, lBalance1, uint104(lReserve0), uint104(lReserve1), lBlockTimestampLast);
         emit Swap(msg.sender, tokenOut == token1, lReceived, amountOut, to);
     }
